Extract downloading-captures functionality to standalone service.
