+++ src/app/shared/services/dia-backend/asset/downloading/dia-backend-downloading.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+import { SharedTestingModule } from '../../../../shared-testing.module';
+import { DiaBackendDownloadingService } from './dia-backend-downloading.service';
+
+describe('DiaBackendDownloadingService', () => {
+  let service: DiaBackendDownloadingService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({ imports: [SharedTestingModule] });
+    service = TestBed.inject(DiaBackendDownloadingService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
