+++ src/app/shared/services/dia-backend/asset/prefetching/dia-backend-asset-prefetching.service.ts
@@ -1,27 +1,14 @@
 import { Injectable } from '@angular/core';
-import { first } from 'rxjs/operators';
-import { blobToBase64 } from '../../../../../utils/encoding/encoding';
-import { OnConflictStrategy } from '../../../database/table/table';
-import { ImageStore } from '../../../image-store/image-store.service';
-import {
-  getSignatures,
-  getTruth,
-} from '../../../repositories/proof/old-proof-adapter';
-import { Proof } from '../../../repositories/proof/proof';
-import { ProofRepository } from '../../../repositories/proof/proof-repository.service';
-import {
-  DiaBackendAsset,
-  DiaBackendAssetRepository,
-} from '../dia-backend-asset-repository.service';
+import { DiaBackendAssetRepository } from '../dia-backend-asset-repository.service';
+import { DiaBackendDownloadingService } from '../downloading/dia-backend-downloading.service';
 
 @Injectable({
   providedIn: 'root',
 })
 export class DiaBackendAssetPrefetchingService {
   constructor(
-    private readonly imageStore: ImageStore,
-    private readonly proofRepository: ProofRepository,
-    private readonly diaBackendAssetRepository: DiaBackendAssetRepository
+    private readonly assetRepository: DiaBackendAssetRepository,
+    private readonly downloadingService: DiaBackendDownloadingService
   ) {}
 
   async prefetch(
@@ -34,7 +21,7 @@ export class DiaBackendAssetPrefetchingService {
       const {
         results: diaBackendAssets,
         count: totalCount,
-      } = await this.diaBackendAssetRepository
+      } = await this.assetRepository
         .fetchCaptures$({ offset: currentOffset, limit })
         .toPromise();
 
@@ -43,52 +30,12 @@ export class DiaBackendAssetPrefetchingService {
       }
       await Promise.all(
         diaBackendAssets.map(async diaBackendAsset => {
-          await this.storeAssetThumbnail(diaBackendAsset);
-          await this.storeIndexedProof(diaBackendAsset);
+          await this.downloadingService.storeRemoteCapture(diaBackendAsset);
           currentCount += 1;
           onStored(currentCount, totalCount);
         })
       );
       currentOffset += diaBackendAssets.length;
     }
-  }
-
-  private async storeAssetThumbnail(diaBackendAsset: DiaBackendAsset) {
-    if (!diaBackendAsset.information.proof) {
-      return;
-    }
-    const thumbnailBlob = await this.diaBackendAssetRepository
-      .downloadFile$({ id: diaBackendAsset.id, field: 'asset_file_thumbnail' })
-      .pipe(first())
-      .toPromise();
-    return this.imageStore.storeThumbnail(
-      diaBackendAsset.proof_hash,
-      await blobToBase64(thumbnailBlob),
-      diaBackendAsset.information.proof.mimeType
-    );
-  }
-
-  private async storeIndexedProof(diaBackendAsset: DiaBackendAsset) {
-    if (
-      !diaBackendAsset.information.proof ||
-      !diaBackendAsset.information.information
-    ) {
-      return;
-    }
-    const proof = new Proof(
-      this.imageStore,
-      getTruth({
-        proof: diaBackendAsset.information.proof,
-        information: diaBackendAsset.information.information,
-      }),
-      getSignatures(diaBackendAsset.signature)
-    );
-    proof.setIndexedAssets({
-      [diaBackendAsset.proof_hash]: {
-        mimeType: diaBackendAsset.information.proof.mimeType,
-      },
-    });
-    proof.diaBackendAssetId = diaBackendAsset.id;
-    return this.proofRepository.add(proof, OnConflictStrategy.REPLACE);
   }
 }
